# docs/swagger.yaml
basePath: /
definitions:
...
  pkg_http.APIResponse:
    properties:
      status:
        type: string
    type: object
  pkg_http.ApproveRequest:
    properties:
      action:
        description: defaults to approve
        type: string
      id:
        type: string
      identifier:
        type: string
      voter:
        type: string
    type: object
  pkg_http.AuditLogsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/types.AuditLog'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  pkg_http.EmptyResponse:
    type: object
...
  pkg_http.ResourceApprovalsUpdateRequest:
    properties:
      identifier:
        type: string
      provider:
        type: string
      votesRequired:
        type: integer
    type: object
...
  types.Approval:
    properties:
      archived:
        description: Archived is set to true once approval is finally approved/rejected
        type: boolean
      createdAt:
        description: When this approval was created
        type: string
      currentVersion:
        type: string
      deadline:
        description: Deadline for this request
        type: string
      digest:
        description: |-
          Digest is used to verify that images are the ones that got the approvals.
          If digest doesn't match for the image, votes are reset.
        type: string
      event:
        allOf:
        - $ref: '#/definitions/types.Event'
        description: Event that triggered evaluation
      id:
        type: string
      identifier:
        description: |-
          Identifier is used to inform user about specific
          Helm release or k8s deployment
          ie: k8s <namespace>/<deployment name>
              helm: <namespace>/<release name>
        type: string
      message:
        type: string
      newVersion:
        type: string
      provider:
        allOf:
        - $ref: '#/definitions/types.ProviderType'
        description: Provider name - Kubernetes/Helm
      rejected:
        description: |-
          Explicitly rejected approval
          can be set directly by user
          so even if deadline is not reached approval
          could be turned down
        type: boolean
      updatedAt:
        description: WHen this approval was updated
        type: string
      voters:
        allOf:
        - $ref: '#/definitions/types.JSONB'
        description: |-
          Voters is a list of voter
          IDs for audit
      votesReceived:
        type: integer
      votesRequired:
        description: |-
          Requirements for the update such as number of votes
          and deadline
        type: integer
    type: object
...
info:
  contact: {}
  description: HTTP API exposed by Keel. Admin routes are registered only when an
    authenticator is enabled. Provider webhooks require Basic or Bearer authorization
    only when authenticated webhooks are enabled; the registry webhook is always unauthenticated.
  title: Keel HTTP API
  version: "1.0"
paths:
...
  /v1/approvals:
    get:
      description: Lists active and archived approvals. This route exists only when
        the authenticator is enabled. Legacy store and serialization error paths write
        text with status 200.
      operationId: listApprovals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Approval'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List approvals
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Approves, rejects, deletes, or archives an approval. Delete returns
        JSON null; legacy delete/archive error paths write text with status 200. This
        route exists only when the authenticator is enabled.
      operationId: updateApproval
      parameters:
      - description: Approval action
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.ApproveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Approval'
        "400":
          description: Malformed request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Approval not found
          schema:
            type: string
        "500":
          description: Approval update failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update an approval
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Sets a Kubernetes resource's required vote count. This route exists
        only when the authenticator is enabled.
      operationId: setResourceApprovals
      parameters:
      - description: Approval requirement
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.ResourceApprovalsUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.APIResponse'
        "400":
          description: Malformed or unsupported request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Permission denied
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Set required approvals
      tags:
      - Admin
  /v1/audit:
    get:
      description: Lists audit entries. Invalid limit or offset values are silently
        treated as zero. This route exists only when the authenticator is enabled.
      operationId: listAuditLogs
      parameters:
      - description: Maximum entries
        in: query
        name: limit
        type: integer
      - description: Entry offset
        in: query
        name: offset
        type: integer
      - description: Comma-separated resource kinds
        in: query
        name: filter
        type: string
      - description: Account email
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.AuditLogsResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Permission denied
          schema:
            type: string
        "500":
          description: Store query failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List audit logs
      tags:
      - Admin
  /v1/auth/info:
    get:
      description: Returns the current administrator profile. This route exists only
        when the authenticator is enabled.
      operationId: getAuthInfo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.UserInfo'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get authentication info
      tags:
      - Auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates Basic credentials supplied as JSON. This route exists
        only when the authenticator is enabled.
      operationId: login
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer token
              type: string
          schema:
            $ref: '#/definitions/github_com_keel-hq_keel_pkg_auth.AuthResponse'
        "400":
          description: Malformed request
          schema:
            type: string
        "401":
          description: Incorrect username or password
          schema:
            type: string
      summary: Log in
      tags:
      - Auth
  /v1/auth/logout:
    get:
      description: Returns an empty JSON object. This route exists only when the authenticator
        is enabled.
      operationId: logoutViaGet
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.EmptyResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Log out with GET
      tags:
      - Auth
    post:
      description: Returns an empty JSON object. This route exists only when the authenticator
        is enabled.
      operationId: logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.EmptyResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Log out
      tags:
      - Auth
  /v1/auth/refresh:
    get:
      description: Generates a new bearer token. This route exists only when the authenticator
        is enabled.
      operationId: refreshAuth
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Refreshed bearer token
              type: string
          schema:
            $ref: '#/definitions/github_com_keel-hq_keel_pkg_auth.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Token generation failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Refresh authentication
      tags:
      - Auth
  /v1/auth/user:
    get:
      description: Returns the current administrator profile. This route exists only
        when the authenticator is enabled.
      operationId: getAuthUser
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.UserInfo'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
  /v1/policies:
    put:
      consumes:
      - application/json
      description: Updates the policy annotation for a Kubernetes resource. This route
        exists only when the authenticator is enabled.
      operationId: updateResourcePolicy
      parameters:
      - description: Policy update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.ResourcePolicyUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.APIResponse'
        "400":
          description: Malformed request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Permission denied
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update resource policy
      tags:
      - Admin
  /v1/resources:
    get:
      description: Returns monitored Kubernetes resources, or JSON null when the source
        slice is nil. This route exists only when the authenticator is enabled.
      operationId: listResources
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pkg_http.ResourceResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List resources
      tags:
      - Admin
  /v1/stats:
    get:
      description: Returns daily webhook, approval, rejection, and update counts.
        This route exists only when the authenticator is enabled.
      operationId: getStats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.AuditLogStats'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Permission denied
          schema:
            type: string
        "500":
          description: Store query failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Get audit statistics
      tags:
      - Admin
  /v1/tracked:
    get:
      description: Returns image polling configuration, or JSON null when the source
        slice is nil. This route exists only when the authenticator is enabled.
      operationId: listTrackedImages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/pkg_http.TrackedImage'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Permission denied
          schema:
            type: string
        "500":
          description: Provider lookup failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: List tracked images
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Sets the trigger and polling schedule for a Kubernetes resource.
        This route exists only when the authenticator is enabled.
      operationId: updateTrackedImage
      parameters:
      - description: Tracking update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.TrackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_http.APIResponse'
        "400":
          description: Malformed or unsupported request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Permission denied
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Update image tracking
      tags:
      - Admin
  /v1/webhooks/azure:
    post:
      consumes:
      - application/json
      description: Requires Basic or Bearer authorization only when authenticatedWebhooks
        is enabled.
      operationId: receiveAzureWebhook
      parameters:
      - description: Azure Container Registry push
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.AzureWebhook'
      responses:
        "200":
          description: Accepted
        "400":
          description: Malformed payload or missing tag
          schema:
            type: string
        "401":
          description: Unauthorized when authenticated webhooks are enabled
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Receive an Azure webhook
      tags:
      - Webhooks
  /v1/webhooks/dockerhub:
    post:
      consumes:
      - application/json
      description: Requires Basic or Bearer authorization only when authenticatedWebhooks
        is enabled.
      operationId: receiveDockerHubWebhook
      parameters:
      - description: Docker Hub push
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.DockerHubWebhook'
      responses:
        "200":
          description: Accepted
        "400":
          description: Malformed payload, missing repository name, or missing tag
          schema:
            type: string
        "401":
          description: Unauthorized when authenticated webhooks are enabled
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Receive a Docker Hub webhook
      tags:
      - Webhooks
  /v1/webhooks/github:
    post:
      consumes:
      - application/json
      description: X-GitHub-Event selects package or registry_package decoding; absent
        and other values currently submit an empty event and return 200. Requires
        Basic or Bearer authorization only when authenticatedWebhooks is enabled.
      operationId: receiveGitHubWebhook
      parameters:
      - description: 'Payload type: package or registry_package'
        in: header
        name: X-GitHub-Event
        type: string
      - description: GitHub Packages or GHCR push
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.GitHubWebhook'
      responses:
        "200":
          description: Accepted
        "400":
          description: Malformed or incomplete supported payload
          schema:
            type: string
        "401":
          description: Unauthorized when authenticated webhooks are enabled
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Receive a GitHub webhook
      tags:
      - Webhooks
  /v1/webhooks/harbor:
    post:
      consumes:
      - application/json
      description: Processes pushImage and PUSH_ARTIFACT resources. Other event types
        return 200 without submitting an event. Requires Basic or Bearer authorization
        only when authenticatedWebhooks is enabled.
      operationId: receiveHarborWebhook
      parameters:
      - description: Harbor artifact push
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.HarborWebhook'
      responses:
        "200":
          description: Accepted
        "400":
          description: Malformed payload or resource URL
          schema:
            type: string
        "401":
          description: Unauthorized when authenticated webhooks are enabled
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Receive a Harbor webhook
      tags:
      - Webhooks
  /v1/webhooks/jfrog:
    post:
      consumes:
      - application/json
      description: Requires Basic or Bearer authorization only when authenticatedWebhooks
        is enabled.
      operationId: receiveJFrogWebhook
      parameters:
      - description: JFrog Docker push
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.JFrogWebhook'
      responses:
        "200":
          description: Accepted
        "400":
          description: Malformed payload, missing image name, or missing tag
          schema:
            type: string
        "401":
          description: Unauthorized when authenticated webhooks are enabled
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Receive a JFrog webhook
      tags:
      - Webhooks
  /v1/webhooks/native:
    post:
      consumes:
      - application/json
      description: Requires Basic or Bearer authorization only when authenticatedWebhooks
        is enabled.
      operationId: receiveNativeWebhook
      parameters:
      - description: Repository event
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.Repository'
      responses:
        "200":
          description: Accepted
        "400":
          description: Malformed repository, missing name, or missing tag
          schema:
            type: string
        "401":
          description: Unauthorized when authenticated webhooks are enabled
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Receive a native webhook
      tags:
      - Webhooks
  /v1/webhooks/quay:
    post:
      consumes:
      - application/json
      description: Requires Basic or Bearer authorization only when authenticatedWebhooks
        is enabled.
      operationId: receiveQuayWebhook
      parameters:
      - description: Quay repository push
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.QuayWebhook'
      responses:
        "200":
          description: Accepted
        "400":
          description: Malformed payload, missing Docker URL, or no updated tags
          schema:
            type: string
        "401":
          description: Unauthorized when authenticated webhooks are enabled
          schema:
            type: string
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Receive a Quay webhook
      tags:
      - Webhooks
  /v1/webhooks/registry:
    post:
      consumes:
      - application/json
      description: Processes push events with tags. This endpoint is always unauthenticated,
        including when authenticatedWebhooks is enabled.
      operationId: receiveRegistryWebhook
      parameters:
      - description: Docker Registry notification
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/pkg_http.RegistryNotification'
      responses:
        "200":
          description: Accepted
        "400":
          description: Malformed payload
      summary: Receive a Docker Registry webhook
      tags:
      - Webhooks
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"